+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Added
+- front: Add possibility to display cores allocations graphical representation
+  in resources page (#712).
 - docs:
   - Add procedure to install Slurm-web on SLES (and openSUSE Leap) 15 and 16 in
     quickstart guide and installation guide (#684).